Remove references to System.Reflection from the ApiListing model
