+++ src/Microsoft.AspNetCore.BuildTools.ApiCheck/ApiListing/ApiListing.cs
@@ -4,8 +4,6 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Reflection;
-using Newtonsoft.Json;
 
 namespace ApiCheck.Description
 {
@@ -14,9 +12,6 @@ namespace ApiCheck.Description
         public string AssemblyIdentity { get; set; }
 
         public IList<TypeDescriptor> Types { get; } = new List<TypeDescriptor>();
-
-        [JsonIgnore]
-        public IEnumerable<Func<MemberInfo, bool>> SourceFilters { get; set; }
 
         public TypeDescriptor FindType(string name)
         {
