+++ src/Microsoft.AspNetCore.BuildTools.ApiCheck/Microsoft.AspNetCore.BuildTools.ApiCheck.csproj
@@ -1,20 +1,27 @@
-﻿<Project Sdk="Microsoft.NET.Sdk" ToolsVersion="15.0">
+﻿<Project Sdk="Microsoft.NET.Sdk">
+  
   <Import Project="..\..\build\common.props" />
+  
   <PropertyGroup>
     <VersionPrefix>1.0.1</VersionPrefix>
     <TargetFrameworks>netcoreapp1.1;net452</TargetFrameworks>
     <OutputType>exe</OutputType>
+    <RootNamespace>ApiCheck</RootNamespace>
   </PropertyGroup>
+  
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.CommandLineUtils" Version="1.0.0" />
     <PackageReference Include="Newtonsoft.Json" Version="$(JsonNetVersion)" />
   </ItemGroup>
+  
   <ItemGroup Condition=" '$(TargetFramework)' == 'netcoreapp1.1' ">
     <PackageReference Include="System.Runtime.Loader" Version="4.3.0" />
   </ItemGroup>
+  
   <ItemGroup>
     <ProjectReference Include="..\NugetReferenceResolver\NugetReferenceResolver.csproj" />
   </ItemGroup>
+  
   <!-- packaging settings-->
   <PropertyGroup>
     <NoPackageAnalysis>true</NoPackageAnalysis>
